refractor to make logic more testable
