+++ src/lib/server/api/people/people.ts
@@ -11,6 +11,11 @@ import { getUniqueKeys } from '$lib/utils/objects/get_unique_keys';
 import { parse, v, mediumString, longString } from '$lib/schema/valibot';
 import { whatsappNumberForVerification } from '$lib/schema/people/channels/channels';
 import type { WhatsappInboundMessage } from '$lib/schema/communications/whatsapp/webhooks/ycloud';
+import * as csv from 'fast-csv';
+import { type Read as ReadInstance } from '$lib/schema/core/instance';
+import { create as createSchema } from '$lib/schema/people/people';
+import { update as updateImport } from '$lib/server/api/people/imports';
+import { getParseSchema, type ImportResult } from '$lib/schema/people/imports';
 
 export const redisString = (instance_id: number, person_id: number | 'all') =>
 	`i:${instance_id}:people:${person_id}`;
@@ -667,4 +672,175 @@ export async function _getPersonByEmail({
 		);
 	}
 	return await read({ instance_id: instanceId, person_id: person[0].id });
+}
+
+export async function parseImportCsv(
+	csvString: string,
+	importId: number,
+	instance: ReadInstance,
+	adminId: number,
+	queue: App.Queue,
+	t: App.Localization
+): Promise<ImportResult> {
+	const records: unknown[] = [];
+	return new Promise((resolve, reject) => {
+		csv
+			.parseString(csvString, { headers: true })
+			.on('error', function (err) {
+				log.error(err);
+			})
+			.on('data', function (row) {
+				// Check if all values in the row are empty or null
+				const isEntirelyEmptyRow = Object.values(row).every(
+					(value) => value === null || String(value).trim() === ''
+				);
+
+				// Check if both family_name and given_name are empty
+				const familyName = row.family_name ? String(row.family_name).trim() : '';
+				const givenName = row.given_name ? String(row.given_name).trim() : '';
+				const isNameEmpty = familyName === '' && givenName === '';
+
+				if (!isEntirelyEmptyRow && !isNameEmpty) {
+					records.push(row);
+				}
+			})
+			.on('end', async function (rowCount: number) {
+				log.debug(`Parsed ${rowCount} rows`);
+				let successCount = 0;
+				let failedCount = 0;
+
+				for (let i = 0; i < records.length; i++) {
+					const parsed = parse(getParseSchema(instance), records[i]);
+					log.debug(parsed);
+					try {
+						//todo: transform phoneNumber, etc into the budgets and add them to the created entries
+						const parsedItem = v.parse(v.looseObject({ ...createSchema.entries }), parsed); //because we want to allow custom fields to be passed through to the function
+						const { events, tags, ...strippedPerson } = parsedItem; //remove events and tags from the person object
+						const createdPerson = await create({
+							instance_id: instance.id,
+							admin_id: adminId,
+							body: strippedPerson,
+							queue: queue,
+							method: 'import'
+						});
+						for (const tag of parsed.tags) {
+							await addPersonToTag({
+								instanceId: instance.id,
+								personId: createdPerson.id,
+								tagName: tag,
+								t
+							});
+						}
+						for (const eventSlug of parsed.events) {
+							await addPersonToEvent({
+								instanceId: instance.id,
+								personId: createdPerson.id,
+								eventSlug,
+								t,
+								queue,
+								importId
+							});
+						}
+						successCount++;
+					} catch (err) {
+						log.error(err);
+						failedCount++;
+					}
+				}
+
+				//await update the import record with the success and failed count and the status
+				await updateImport({
+					instanceId: instance.id,
+					importId,
+					t,
+					body: {
+						status: 'complete',
+						total_rows: records.length,
+						processed_rows: successCount,
+						failed_rows: failedCount
+					}
+				});
+
+				resolve({
+					totalRows: records.length,
+					successCount,
+					failedCount,
+					records
+				});
+			});
+	});
+}
+
+import { readBySlug } from '$lib/server/api/events/events';
+import { create as createAttendee } from '$lib/server/api/events/attendees';
+async function addPersonToEvent({
+	instanceId,
+	personId,
+	eventSlug,
+	t,
+	queue,
+	importId
+}: {
+	instanceId: number;
+	personId: number;
+	eventSlug: string;
+	t: App.Localization;
+	queue: App.Queue;
+	importId: number;
+}): Promise<void> {
+	try {
+		const event = await readBySlug({ instanceId, slug: eventSlug });
+		await createAttendee({
+			instanceId,
+			eventId: event.id,
+			body: {
+				person_id: personId,
+				status: 'attended',
+				send_notifications: false,
+				notes: `import:${importId}`,
+				response_channel: 'none'
+			},
+			t,
+			queue
+		});
+		log.debug({ instanceId, personId, eventSlug }, 'Added person to event');
+	} catch (err) {
+		//no event?
+		log.debug(
+			{ instanceId, personId, eventSlug },
+			'Unable to add person to event. Probably no event exists with this slug'
+		);
+	}
+}
+
+import { readByName, create as createTag } from '$lib/server/api/core/tags';
+import { create as createTagging } from '$lib/server/api/people/taggings';
+async function addPersonToTag({
+	instanceId,
+	personId,
+	tagName,
+	t
+}: {
+	instanceId: number;
+	personId: number;
+	tagName: string;
+	t: App.Localization;
+}): Promise<void> {
+	try {
+		//try to add them to an existing tag
+		const tag = await readByName({ instanceId, tagName });
+		await createTagging({ instanceId, personId, tagId: tag.id, t });
+		log.debug({ instanceId, personId, tagName }, 'Added tag');
+	} catch (err) {
+		//if no tag exists, readByName will error. So we create a new tag
+		try {
+			const newTag = await createTag({ instanceId, body: { name: tagName } });
+			log.debug(newTag, 'Created new tag');
+			await createTagging({ instanceId, personId, tagId: newTag.id, t });
+			log.debug({ instanceId, personId, tagName, newTag }, 'Added person to new tag');
+		} catch (err) {
+			//If something went wrong with the creation or adding of a new tag.
+			log.debug({ instanceId, personId, tagName }, 'Unable to create tag');
+		}
+	}
 }
