+++ src/lib/utils/mocks/instance.mock.ts
@@ -23,7 +23,8 @@ export const events = {
 export const communications = {
 	email: {
 		default_from_name: 'Test',
-		default_template_id: 3
+		default_template_id: 3,
+		default_template_name: 'main' as const
 	},
 	whatsapp: {
 		default_template_id: 4,
@@ -50,10 +51,14 @@ export const website = {
 export const settings = {
 	default_admin_id: 1,
 	home_page_url: 'localhost',
-	events,
+	events: {
+		...events,
+		event_email_template_prefix: 'test-'
+	},
 	communications,
 	petitions,
-	website
+	website,
+	default_timezone: 'Europe/London'
 };
 
 export const instance: Read = {
